requirements/local.txt
# including production so we can run foreman
-r production.txt
# already includes base.txt

coverage==3.6
django-discover-runner==0.4
django-debug-toolbar==0.9.4
Sphinx==1.2b1
